Changed on/off times to uint16

+++ fsw/src/gpio_ctrl.h
@@ -105,8 +105,8 @@ typedef struct
    bool    IsMapped;
    bool    LedOn;
    uint8   OutPin;
-   uint32  OnTime;    /* Time in Milliseconds */
-   uint32  OffTime;   /* Time in Milliseconds */ 
+   uint16  OnTime;    /* Time in Milliseconds */
+   uint16  OffTime;   /* Time in Milliseconds */ 
    
    /*
    ** Telemetry Packets
